ModelTrainReport/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "matplotlib>=3.10.6",
    "pydantic>=2.11.7",
    "scikit-learn>=1.7.2",
    "streamlit-ace>=0.1.1",
    "streamlit>=1.49.1",
    "tensorflow>=2.20.0",
    "numpy>=2.3.3",
    "plotly>=6.3.0",
    "weasyprint>=66.0",
    "onnxruntime>=1.22.1",
    "mlflow>=3.3.2",
    "pandas>=2.3.2",
    "jinja2>=3.1.6",
    "pillow>=11.3.0",
    "opencv-python>=4.11.0.86",
    "uvicorn>=0.35.0",
    "seaborn>=0.13.2",
    "fastapi>=0.116.1",
    "scipy>=1.16.1",
    "tf2onnx>=1.8.4",
]
